+++ game.json
@@ -20,6 +20,16 @@
 			"width": 100,
 			"height": 100,
 			"path": "image/circleB.png"
+		},
+		"parameters": {
+			"type": "script",
+			"path": "script/parameters.js",
+			"global": true
+		},
+		"utils": {
+			"type": "script",
+			"path": "script/utils.js",
+			"global": true
 		}
 	},
 	"environment": {
